Refactor, move to directories
