+++ apps/app/lib/app/di/bloc_providers_factory.dart
@@ -3,9 +3,9 @@ import 'package:anyhoo_logging/anyhoo_logging.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:core_models/core_models.dart';
 import 'package:listwhatever/app/features/list_items/cubit/list_items_cubit.dart';
-import 'package:listwhatever/app/features/list_items/list_items_repository.dart';
+import 'package:listwhatever/app/features/list_items/repositories/list_items_repository.dart';
 import 'package:listwhatever/app/features/lists/cubit/dynamic_lists_cubit.dart';
-import 'package:listwhatever/app/features/lists/dynamic_lists_repository.dart';
+import 'package:listwhatever/app/features/lists/repositories/dynamic_lists_repository.dart';
 
 abstract final class BlocProvidersFactory {
   /// App-wide cubits. Use [BlocProvider.create] and [BuildContext.read] for
