+++ apps/app/lib/app/features/list_items/list_map_view.dart
@@ -2,8 +2,8 @@ import 'package:anyhoo_design_system/anyhoo_design_system.dart';
 import 'package:core_models/core_models.dart';
 import 'package:flutter/material.dart';
 import 'package:listwhatever/app/features/list_items/list_items_view.dart';
-import 'package:listwhatever/app/features/list_items/list_map_placeholder.dart';
-import 'package:listwhatever/app/features/list_items/map_item_peek_sheet.dart';
+import 'package:listwhatever/app/features/list_items/widgets/list_map_placeholder.dart';
+import 'package:listwhatever/app/features/list_items/widgets/map_item_peek_sheet.dart';
 import 'package:listwhatever/i18n/strings.g.dart';
 
 class ListMapView extends StatelessWidget {
