+++ apps/app/lib/app/features/list_items/widgets/list_items_skeleton.dart
