+++ apps/app/lib/app/main/app.dart
@@ -7,8 +7,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 import 'package:listwhatever/app/di/bloc_providers_factory.dart';
-import 'package:listwhatever/app/features/list_items/list_items_repository.dart';
-import 'package:listwhatever/app/features/lists/dynamic_lists_repository.dart';
+import 'package:listwhatever/app/features/list_items/repositories/list_items_repository.dart';
+import 'package:listwhatever/app/features/lists/repositories/dynamic_lists_repository.dart';
 import 'package:listwhatever/app/main/app_theme.dart';
 import 'package:listwhatever/app/main/environments/interfaces/app_repositories.dart';
 import 'package:listwhatever/app/main/environments/prod/flavors/flavor_profile.dart';
