+++ apps/app/lib/app/main/environments/fake/fake_repositories.dart
@@ -1,8 +1,8 @@
 import 'package:anyhoo_core/repositories/current_time_repository.dart';
-import 'package:listwhatever/app/features/list_items/fake_list_items_repository.dart';
-import 'package:listwhatever/app/features/list_items/list_items_repository.dart';
-import 'package:listwhatever/app/features/lists/dynamic_lists_repository.dart';
-import 'package:listwhatever/app/features/lists/fake_dynamic_lists_repository.dart';
+import 'package:listwhatever/app/features/list_items/repositories/fake_list_items_repository.dart';
+import 'package:listwhatever/app/features/list_items/repositories/list_items_repository.dart';
+import 'package:listwhatever/app/features/lists/repositories/dynamic_lists_repository.dart';
+import 'package:listwhatever/app/features/lists/repositories/fake_dynamic_lists_repository.dart';
 import 'package:listwhatever/app/main/environments/interfaces/app_repositories.dart';
 
 class FakeRepositories extends AppRepositories {
