+++ apps/app/lib/app/main/environments/interfaces/app_repositories.dart
@@ -1,6 +1,6 @@
 import 'package:anyhoo_core/repositories/current_time_repository.dart';
-import 'package:listwhatever/app/features/list_items/list_items_repository.dart';
-import 'package:listwhatever/app/features/lists/dynamic_lists_repository.dart';
+import 'package:listwhatever/app/features/list_items/repositories/list_items_repository.dart';
+import 'package:listwhatever/app/features/lists/repositories/dynamic_lists_repository.dart';
 
 abstract class AppRepositories {
   DynamicListsRepository get dynamicListsRepository;
