+++ apps/app/lib/app/main/environments/prod/production_repositories.dart
@@ -1,9 +1,9 @@
 import 'package:anyhoo_firebase/anyhoo_firebase.dart';
-import 'package:listwhatever/app/features/list_items/list_items_repository.dart';
-import 'package:listwhatever/app/features/lists/dynamic_lists_repository.dart';
+import 'package:listwhatever/app/features/list_items/repositories/list_items_repository.dart';
+import 'package:listwhatever/app/features/lists/repositories/dynamic_lists_repository.dart';
 import 'package:listwhatever/app/main/environments/interfaces/app_repositories.dart';
-import 'package:listwhatever/app/main/environments/prod/firestore_dynamic_lists_repository.dart';
-import 'package:listwhatever/app/main/environments/prod/firestore_list_items_repository.dart';
+import 'package:listwhatever/app/features/lists/repositories/firestore_dynamic_lists_repository.dart';
+import 'package:listwhatever/app/features/list_items/repositories/firestore_list_items_repository.dart';
 
 class ProductionRepositories extends AppRepositories {
   ProductionRepositories({required this.firestoreService});
