+++ apps/app/test/goldens/list_items_page_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:listwhatever/app/features/list_items/cubit/list_items_cubit.dart';
-import 'package:listwhatever/app/features/list_items/fake_list_items_repository.dart';
+import 'package:listwhatever/app/features/list_items/repositories/fake_list_items_repository.dart';
 import 'package:listwhatever/app/features/list_items/list_items_page.dart';
 import 'package:listwhatever/app/features/list_items/list_items_view.dart';
